Implements functionality to duplicate the last run drawing
This commit implements the "duplicate" button on the administrator's
drawings view. When this button is pressed, the user is asked to
specify the date of the next drawing and then the duplicated drawing
is scheduled for that date and time. The new drawing will not include
the won items or entrants from it's parent.